requirements.txt
asgiref==3.3.1
autopep8==1.5.6
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.2
gcloud==0.17.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.0
jws==0.1.3
oauth2client==3.0.0
protobuf==3.15.6
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycryptodome==3.4.3
pyparsing==2.4.7
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2021.1
requests==2.11.1
requests-toolbelt==0.7.0
rope==0.18.0
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0